+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-file-view/components/case-file-view-folder/case-file-view-folder.component.spec.ts
@@ -1,5 +1,6 @@
 import { CdkTreeModule } from '@angular/cdk/tree';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ReactiveFormsModule } from '@angular/forms';
 import { RouterTestingModule } from '@angular/router/testing';
 import createSpyObj = jasmine.createSpyObj;
 import SpyObj = jasmine.SpyObj;
@@ -8,10 +9,11 @@ import { of } from 'rxjs';
 import {
   CaseFileViewDocument,
   DocumentTreeNode
+  DocumentTreeNodeType
 } from '../../../../../domain/case-file-view';
 import { DocumentManagementService, WindowService } from '../../../../../services';
 import { categoriesAndDocuments } from '../../test-data/categories-and-documents-test-data';
-import { treeData } from '../../test-data/document-tree-node-test-data';
+import { treeData, treeDataWithUncategorisedDocuments } from '../../test-data/document-tree-node-test-data';
 import { CaseFileViewFolderComponent, MEDIA_VIEWER_LOCALSTORAGE_KEY } from './case-file-view-folder.component';
 
 describe('CaseFileViewFolderComponent', () => {
@@ -36,7 +38,7 @@ describe('CaseFileViewFolderComponent', () => {
     TestBed.configureTestingModule({
       imports: [
         CdkTreeModule,
-        RouterTestingModule
+        ReactiveFormsModule
       ],
       declarations: [
         CaseFileViewFolderComponent
@@ -55,8 +57,17 @@ describe('CaseFileViewFolderComponent', () => {
     fixture.detectChanges();
   }));
 
-  it('should create', () => {
-    expect(component).toBeTruthy();
+  it('should create', async() => {
+    spyOn(component, 'filter').and.returnValue(of([]));
+    const documentFilterInputEl = nativeElement.querySelector('.document-search');
+    documentFilterInputEl.dispatchEvent(new Event('focusin'));
+    documentFilterInputEl.value = 'enc';
+    documentFilterInputEl.dispatchEvent(new Event('input'));
+    fixture.detectChanges();
+    await fixture.whenStable();
+    fixture.detectChanges();
+    expect(component.filter).toHaveBeenCalled();
+    expect(component.documentTreeData).toEqual(treeDataWithUncategorisedDocuments);
   });
 
   it('should generate tree data', () => {
@@ -68,19 +79,19 @@ describe('CaseFileViewFolderComponent', () => {
     const documentsTreeNodes: DocumentTreeNode[] = plainToClass(DocumentTreeNode, [
       {
         name: 'Lager encyclopedia',
-        type: 'document',
+        type: DocumentTreeNodeType.DOCUMENT,
         document_filename: 'Lager encyclopedia',
         document_binary_url: '/test/binary'
       },
       {
         name: 'Beers encyclopedia',
-        type: 'document',
+        type: DocumentTreeNodeType.DOCUMENT,
         document_filename: 'Beers encyclopedia',
         document_binary_url: '/test/binary'
       },
       {
         name: 'Ale encyclopedia',
-        type: 'document',
+        type: DocumentTreeNodeType.DOCUMENT,
         document_filename: 'Ale encyclopedia',
         document_binary_url: '/test/binary'
       }
@@ -108,17 +119,17 @@ describe('CaseFileViewFolderComponent', () => {
     ];
     const uncategorisedDocumentsTreeNode: DocumentTreeNode = plainToClass(DocumentTreeNode, {
       name: 'Uncategorised documents',
-      type: 'category',
+      type: DocumentTreeNodeType.FOLDER,
       children: [
         {
           name: 'Uncategorised document 1',
-          type: 'document',
+          type: DocumentTreeNodeType.DOCUMENT,
           document_filename: 'Uncategorised document 1',
           document_binary_url: '/test/binary',
         },
         {
           name: 'Uncategorised document 2',
-          type: 'document',
+          type: DocumentTreeNodeType.DOCUMENT,
           document_filename: 'Uncategorised document 2',
           document_binary_url: '/test/binary',
         }
@@ -130,7 +141,7 @@ describe('CaseFileViewFolderComponent', () => {
   it('should render cdk nested tree', () => {
     component.nestedDataSource = treeData;
     fixture.detectChanges();
-    const documentTreeContainerEl = nativeElement.querySelector('.document-tree-container');
+    const documentTreeContainerEl = nativeElement.querySelector('.document-tree');
     expect(documentTreeContainerEl).toBeDefined();
   });
 
@@ -175,9 +186,156 @@ describe('CaseFileViewFolderComponent', () => {
     expect(component.windowService.openOnNewTab).toHaveBeenCalledWith('/media-viewer');
   });
 
+  it('should display correct folder icons', () => {
+    component.nestedDataSource = treeData;
+    fixture.detectChanges();
+    const documentTreeContainerEl = nativeElement.querySelector('.document-tree');
+    const firstNodeButton = documentTreeContainerEl.querySelector('.node-button');
+    const iconEl = firstNodeButton.querySelector('.icon');
+    expect(iconEl.getAttribute('src')).toEqual('/assets/images/folder.png');
+    firstNodeButton.click();
+    fixture.detectChanges();
+    expect(iconEl.getAttribute('src')).toEqual('/assets/images/folder-open.png');
+  });
+
+  it('should filter documents', () => {
+    const filteredTreeData: DocumentTreeNode[] = [
+      {
+        name: 'Spirits',
+        type: DocumentTreeNodeType.FOLDER,
+        children: [
+          {
+            name: 'Scotch whisky',
+            type: DocumentTreeNodeType.FOLDER,
+            children: [
+              {
+                name: 'Lowland',
+                type: DocumentTreeNodeType.FOLDER,
+                children: [
+                  {
+                    name: 'Lowland 1',
+                    type: DocumentTreeNodeType.FOLDER,
+                    children: [
+                      {
+                        name: 'Details about Whisky Lowland 1',
+                        type: DocumentTreeNodeType.DOCUMENT
+                      }
+                    ]
+                  }
+                ]
+              },
+              {
+                name: 'Islay',
+                type: DocumentTreeNodeType.FOLDER,
+                children: [
+                  {
+                    name: 'Details about Whisky Islay',
+                    type: DocumentTreeNodeType.DOCUMENT
+                  },
+                  {
+                    name: 'More information about Whisky Islay',
+                    type: DocumentTreeNodeType.DOCUMENT
+                  }
+                ]
+              }
+            ]
+          }
+        ]
+      }
+    ];
+    component.documentTreeData = filteredTreeData;
+    component.documentSearchFormControl.setValue('abo');
+    component.filter('abo').subscribe(result => {
+      expect(result).toEqual(filteredTreeData);
+    });
+  });
+
+  it('should filter documents no match', () => {
+    const filteredTreeData: DocumentTreeNode[] = [
+      {
+        name: 'Spirits',
+        type: DocumentTreeNodeType.FOLDER,
+        children: [
+          {
+            name: 'Scotch whisky',
+            type: DocumentTreeNodeType.FOLDER,
+            children: [
+              {
+                name: 'Lowland',
+                type: DocumentTreeNodeType.FOLDER,
+                children: [
+                  {
+                    name: 'Lowland 1',
+                    type: DocumentTreeNodeType.FOLDER,
+                    children: [
+                      {
+                        name: 'Details about Whisky Lowland 1',
+                        type: DocumentTreeNodeType.DOCUMENT
+                      }
+                    ]
+                  }
+                ]
+              },
+              {
+                name: 'Islay',
+                type: DocumentTreeNodeType.FOLDER,
+                children: [
+                  {
+                    name: 'Details about Whisky Islay',
+                    type: DocumentTreeNodeType.DOCUMENT
+                  },
+                  {
+                    name: 'More information about Whisky Islay',
+                    type: DocumentTreeNodeType.DOCUMENT
+                  }
+                ]
+              }
+            ]
+          }
+        ]
+      }
+    ];
+    component.documentTreeData = filteredTreeData;
+    component.documentSearchFormControl.setValue('some random text');
+    component.filter('some random text').subscribe(result => {
+      expect(result.length).toEqual(0);
+    });
+  });
+
+  it('should filter documents verify UI', async() => {
+    component.nestedDataSource = treeData;
+    fixture.detectChanges();
+    const documentFilterInputEl = nativeElement.querySelector('.document-search');
+    documentFilterInputEl.dispatchEvent(new Event('focusin'));
+    documentFilterInputEl.value = 'enc';
+    documentFilterInputEl.dispatchEvent(new Event('input'));
+    fixture.detectChanges();
+    await fixture.whenStable();
+    fixture.detectChanges();
+    const documentTreeContainerEl = nativeElement.querySelector('.document-tree');
+    expect(documentTreeContainerEl.textContent).toContain('Beers encyclopedia');
+  });
+
+  it('should filter documents no match verify UI', async() => {
+    component.nestedDataSource = treeData;
+    fixture.detectChanges();
+    const documentFilterInputEl = nativeElement.querySelector('.document-search');
+    documentFilterInputEl.dispatchEvent(new Event('focusin'));
+    documentFilterInputEl.value = 'some random text';
+    documentFilterInputEl.dispatchEvent(new Event('input'));
+    fixture.detectChanges();
+    await fixture.whenStable();
+    fixture.detectChanges();
+    const documentTreeContainerEl = nativeElement.querySelector('.document-tree');
+    expect(documentTreeContainerEl.textContent).toContain('No results found');
+  });
+
+
   it('should unsubscribe', () => {
     spyOn(component.categoriesAndDocumentsSubscription, 'unsubscribe').and.callThrough();
+    spyOn(component.documentFilterSubscription, 'unsubscribe').and.callThrough();
     component.ngOnDestroy();
     expect(component.categoriesAndDocumentsSubscription.unsubscribe).toHaveBeenCalled();
+    expect(component.documentFilterSubscription.unsubscribe).toHaveBeenCalled();
   });
 });