+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-file-view/components/case-file-view-folder/case-file-view-folder.component.ts
@@ -1,10 +1,17 @@
 import { NestedTreeControl } from '@angular/cdk/tree';
 import { Component, Input, OnDestroy, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import { FormControl, FormGroup } from '@angular/forms';
 import { Observable, of, Subscription } from 'rxjs';
-import { CaseFileViewCategory, CaseFileViewDocument, CategoriesAndDocuments, DocumentTreeNode } from '../../../../../domain/case-file-view';
+import { switchMap, tap } from 'rxjs/operators';
+import {
+  CaseFileViewCategory,
+  CaseFileViewDocument,
+  CategoriesAndDocuments,
+  DocumentTreeNode,
+  DocumentTreeNodeType
+} from '../../../../../domain/case-file-view';
 import { DocumentManagementService, WindowService } from '../../../../../services';
-
 export const MEDIA_VIEWER_LOCALSTORAGE_KEY = 'media-viewer-info';
 
 @Component({
@@ -14,6 +21,8 @@ export const MEDIA_VIEWER_LOCALSTORAGE_KEY = 'media-viewer-info';
 })
 export class CaseFileViewFolderComponent implements OnInit, OnDestroy {
   private static readonly UNCATEGORISED_DOCUMENTS_TITLE = 'Uncategorised documents';
+  private static readonly DOCUMENT_SEARCH_FORM_CONTROL_NAME = 'documentSearchFormControl';
+  private static readonly MINIMUM_SEARCH_CHARACTERS = 3;
 
   @Input() public categoriesAndDocuments: Observable<CategoriesAndDocuments>;
 
@@ -21,6 +30,11 @@ export class CaseFileViewFolderComponent implements OnInit, OnDestroy {
   public nestedDataSource: DocumentTreeNode[];
   public categories: CaseFileViewCategory[] = [];
   public categoriesAndDocumentsSubscription: Subscription;
+  public documentFilterFormGroup: FormGroup;
+  public documentSearchFormControl: FormControl;
+  public documentTreeData: DocumentTreeNode[];
+  public documentFilterSubscription: Subscription;
+  public searchTermLength: number;
 
   private getChildren = (node: DocumentTreeNode) => of(node.children);
   public nestedChildren = (_: number, nodeData: DocumentTreeNode) => nodeData.children;
@@ -34,18 +48,35 @@ export class CaseFileViewFolderComponent implements OnInit, OnDestroy {
   }
 
   public ngOnInit(): void {
+    this.documentFilterFormGroup = new FormGroup({});
+    this.documentSearchFormControl = new FormControl('');
+    this.documentFilterFormGroup.addControl(CaseFileViewFolderComponent.DOCUMENT_SEARCH_FORM_CONTROL_NAME, this.documentSearchFormControl);
+
+    // Listen to search input and initiate filter documents if at least three characters entered
+    this.documentFilterSubscription = this.documentSearchFormControl.valueChanges.pipe(
+      tap((searchTerm: string) => this.searchTermLength = searchTerm.length),
+      switchMap((searchTerm: string) => this.filter(searchTerm.toLowerCase()).pipe())
+    ).subscribe(documentTreeData => {
+      this.nestedDataSource = documentTreeData;
+      this.nestedTreeControl.dataNodes = documentTreeData;
+      this.searchTermLength >= CaseFileViewFolderComponent.MINIMUM_SEARCH_CHARACTERS
+        ? this.nestedTreeControl.expandAll()
+        : this.nestedTreeControl.collapseAll();
+    });
+
+    // Subscribe to the input categories and documents, and generate tree data and initialise cdk tree
     this.categoriesAndDocumentsSubscription = this.categoriesAndDocuments.subscribe(categoriesAndDocuments => {
-      // Using the mock data for now as we have to display the documents as well for demo purpose
-      const categories = this.loadCategories(); // categoriesAndDocuments.categories;
+      const categories = categoriesAndDocuments.categories;
       // Generate document tree data from categories
-      const treeData = this.generateTreeData(categories);
+      this.documentTreeData = this.generateTreeData(categories);
       // Append uncategorised documents
       if (categoriesAndDocuments.uncategorised_documents && categoriesAndDocuments.uncategorised_documents.length > 0) {
         const uncategorisedDocuments = this.getUncategorisedDocuments(categoriesAndDocuments.uncategorised_documents);
-        treeData.push(uncategorisedDocuments);
+        this.documentTreeData.push(uncategorisedDocuments);
       }
       // Initialise cdk tree with generated data
-      this.nestedDataSource = treeData;
+      this.nestedDataSource = this.documentTreeData;
+      this.nestedTreeControl.dataNodes = this.documentTreeData;
     });
   }
 
@@ -53,7 +84,7 @@ export class CaseFileViewFolderComponent implements OnInit, OnDestroy {
     return categories.reduce((tree, node) => {
       const newDocumentTreeNode = new DocumentTreeNode();
       newDocumentTreeNode.name = node.category_name;
-      newDocumentTreeNode.type = 'category';
+      type: DocumentTreeNodeType.FOLDER,
       newDocumentTreeNode.children = [...this.generateTreeData(node.sub_categories), ...this.getDocuments(node.documents)];
 
       return [
@@ -68,7 +99,7 @@ export class CaseFileViewFolderComponent implements OnInit, OnDestroy {
     documents.forEach(document => {
       const documentTreeNode = new DocumentTreeNode();
       documentTreeNode.name = document.document_filename;
-      documentTreeNode.type = 'document';
+      documentTreeNode.type = DocumentTreeNodeType.DOCUMENT;
       documentTreeNode.document_filename = document.document_filename;
       documentTreeNode.document_binary_url = document.document_binary_url;
 
@@ -83,7 +114,7 @@ export class CaseFileViewFolderComponent implements OnInit, OnDestroy {
     uncategorisedDocuments.forEach(document => {
       const documentTreeNode = new DocumentTreeNode();
       documentTreeNode.name = document.document_filename;
-      documentTreeNode.type = 'document';
+      documentTreeNode.type = DocumentTreeNodeType.DOCUMENT;
       documentTreeNode.document_filename = document.document_filename;
       documentTreeNode.document_binary_url = document.document_binary_url;
 
@@ -92,7 +123,7 @@ export class CaseFileViewFolderComponent implements OnInit, OnDestroy {
 
     const uncategorisedNode = new DocumentTreeNode();
     uncategorisedNode.name = CaseFileViewFolderComponent.UNCATEGORISED_DOCUMENTS_TITLE;
-    uncategorisedNode.type = 'category';
+    uncategorisedNode.type = DocumentTreeNodeType.FOLDER;
     uncategorisedNode.children = documents;
 
     return uncategorisedNode;
@@ -119,6 +150,27 @@ export class CaseFileViewFolderComponent implements OnInit, OnDestroy {
     });
 
     this.updateNodeData(sortedData);
+  }
+
+  public filter(searchTerm: string): Observable<DocumentTreeNode[]> {
+    // Make a copy of the data so we do not mutate the original
+    function copy(node: DocumentTreeNode) {
+      return Object.assign({}, node);
+    }
+
+    let filteredData = this.documentTreeData;
+    if (searchTerm && searchTerm.length >= CaseFileViewFolderComponent.MINIMUM_SEARCH_CHARACTERS && this.documentFilterFormGroup.controls[CaseFileViewFolderComponent.DOCUMENT_SEARCH_FORM_CONTROL_NAME].value.length > 2) {
+      filteredData = this.documentTreeData.map(copy).filter(function filterTreeData(node: DocumentTreeNode) {
+        if (node.name && node.name.toLowerCase().includes(searchTerm) && node.type === DocumentTreeNodeType.DOCUMENT) {
+          return true;
+        }
+        // Call recursively if node has children
+        if (node.children) {
+          return (node.children = node.children.map(copy).filter(filterTreeData)).length;
+        }
+      });
+    }
+    return of(filteredData);
   }
 
   public triggerDocumentAction(
@@ -155,6 +207,9 @@ export class CaseFileViewFolderComponent implements OnInit, OnDestroy {
     if (this.categoriesAndDocumentsSubscription) {
       this.categoriesAndDocumentsSubscription.unsubscribe();
     }
+    if (this.documentFilterSubscription) {
+      this.documentFilterSubscription.unsubscribe();
+    }
   }
 
   public updateNodeData(data: DocumentTreeNode[]) {
@@ -180,180 +235,5 @@ export class CaseFileViewFolderComponent implements OnInit, OnDestroy {
       return prevSelected.includes(item.name);
     });
     newObjects.forEach(object => this.nestedTreeControl.expand(object));
-  }
-
-  public loadCategories(): CaseFileViewCategory[] {
-    return [
-      {
-        category_id: 'Beers',
-        category_name: 'Beers',
-        category_order: 1,
-        documents: [
-          {
-            document_url: '/test',
-            document_filename: 'Lager encyclopedia',
-            document_binary_url: '/test/binary',
-            attribute_path: '',
-            upload_timestamp: ''
-          },
-          {
-            document_url: '/test',
-            document_filename: 'Beers encyclopedia',
-            document_binary_url: '/test/binary',
-            attribute_path: '',
-            upload_timestamp: ''
-          },
-          {
-            document_url: '/test',
-            document_filename: 'Ale encyclopedia',
-            document_binary_url: '/test/binary',
-            attribute_path: '',
-            upload_timestamp: ''
-          }
-        ],
-        sub_categories: [
-          {
-            category_id: 'BeersBitters',
-            category_name: 'Bitters',
-            category_order: 1,
-            documents: [],
-            sub_categories: []
-          },
-          {
-            category_id: 'BeersAmerican',
-            category_name: 'American',
-            category_order: 2,
-            documents: [],
-            sub_categories: []
-          },
-          {
-            category_id: 'BeersAsian',
-            category_name: 'Asian',
-            category_order: 3,
-            documents: [],
-            sub_categories: []
-          }
-        ]
-      },
-      {
-        category_id: 'Wines',
-        category_name: 'Wines',
-        category_order: 2,
-        documents: [],
-        sub_categories: [
-          {
-            category_id: 'WinesFrench',
-            category_name: 'French',
-            category_order: 1,
-            documents: [],
-            sub_categories: []
-          },
-          {
-            category_id: 'WinesItalian',
-            category_name: 'Italian',
-            category_order: 2,
-            documents: [],
-            sub_categories: []
-          }
-        ]
-      },
-      {
-        category_id: 'Spirits',
-        category_name: 'Spirits',
-        category_order: 3,
-        documents: [],
-        sub_categories: [
-          {
-            category_id: 'SpiritsWhisky',
-            category_name: 'Scotch whisky',
-            category_order: 1,
-            documents: [],
-            sub_categories: [
-              {
-                category_id: 'WhiskyHighland',
-                category_name: 'Highland',
-                category_order: 1,
-                documents: [],
-                sub_categories: [
-                  {
-                    category_id: 'WhiskyHighland1',
-                    category_name: 'Highland 1',
-                    category_order: 1,
-                    documents: [],
-                    sub_categories: []
-                  }
-                ]
-              },
-              {
-                category_id: 'WhiskyLowland',
-                category_name: 'Lowland',
-                category_order: 2,
-                documents: [],
-                sub_categories: [
-                  {
-                    category_id: 'WhiskyLowland1',
-                    category_name: 'Lowland 1',
-                    category_order: 1,
-                    documents: [
-                      {
-                        document_url: '/test',
-                        document_filename: 'Details about Whisky Lowland 1',
-                        document_binary_url: '/test/binary',
-                        attribute_path: '',
-                        upload_timestamp: ''
-                      }
-                    ],
-                    sub_categories: []
-                  },
-                  {
-                    category_id: 'WhiskyLowland2',
-                    category_name: 'Lowland 2',
-                    category_order: 2,
-                    documents: [],
-                    sub_categories: []
-                  }
-                ]
-              },
-              {
-                category_id: 'WhiskyIslay',
-                category_name: 'Islay',
-                category_order: 3,
-                documents: [
-                  {
-                    document_url: '/test',
-                    document_filename: 'Details about Whisky Islay',
-                    document_binary_url: '/test/binary',
-                    attribute_path: '',
-                    upload_timestamp: ''
-                  },
-                  {
-                    document_url: '/test',
-                    document_filename: 'More information about Whisky Islay',
-                    document_binary_url: '/test/binary',
-                    attribute_path: '',
-                    upload_timestamp: ''
-                  }
-                ],
-                sub_categories: []
-              },
-              {
-                category_id: 'WhiskySpeyside',
-                category_name: 'Speyside',
-                category_order: 4,
-                documents: [],
-                sub_categories: []
-              },
-              {
-                category_id: 'WhiskyCampbeltown',
-                category_name: 'Campbeltown',
-                category_order: 5,
-                documents: [],
-                sub_categories: []
-              }
-            ]
-          }
-        ]
-      }
-    ];
   }
 }