+++ projects/ccd-case-ui-toolkit/src/lib/shared/domain/case-file-view/document-tree-node.model.ts
@@ -1,0 +1,11 @@
+export interface DocumentTreeNode {
+  name?: string;
+  type?: DocumentTreeNodeType;
+  count?: number;
+  children?: DocumentTreeNode[];
+}
+
+export enum DocumentTreeNodeType {
+  FOLDER = 'folder',
+  DOCUMENT = 'document',
+}