+++ RELEASE-NOTES.md
@@ -1,4 +1,9 @@
 ## RELEASE NOTES
+### Version 5.0.25-case-file-view-document-tree-functionalities-v3
+**EUI-7066** Case file view search document exceptions
+**EUI-7160** Case file view hand icon for folder clicking
+**EUI-5458** Case file view document view and folder action
+
 ### Version 5.0.24-case-file-view-folder-document-icons-v3
 **EUI-6974** Case file view folder and document icons display